requirements.txt
fire
SQLAlchemy
python-dotenv